Dev #68: adding new implementation of load_thickness
Committed alongside the original implementation for comparison, you can run with examples/config.mem.json to produce the memory allocation issue. The new implementation optimises the generation of the dataframe by utilising xarray alongside generators, which is then used to output the dataframe in it's original structure. This makes everything nicely dask-able in the future (xarray will leverage dask under the hood if a client is available)